many modifications to coin rain and a few other things. -moved coin rain into a skyoceanOverlay. -made the coin rain not framerate dependant (kinda). -added the redstone coin to list of items that can rain. -reorganized stuff a bit. -changed default duration of coins dropping from 5 to 3 -other stuff I probably forgot about (ive been slowly working on this for like 5 days)
